Split map view from SocialLocationPreview
