+++ src/screens/internal/ShareImage/USMapImage.tsx
@@ -1,5 +1,5 @@
 import { SCREENSHOT_CLASS } from 'components/Screenshot';
-import SocialLocationPreview from 'components/SocialLocationPreview/SocialLocationPreview';
+import SocialLocationPreviewMap from 'components/SocialLocationPreview/SocialLocationPreviewMap';
 import React from 'react';
 import { ScreenshotWrapper } from './ShareImage.style';
 import { USMapWrapper } from './USMapImage.style';
@@ -8,7 +8,7 @@ export const USMapImage = () => {
   return (
     <ScreenshotWrapper className={SCREENSHOT_CLASS}>
       <USMapWrapper>
-        <SocialLocationPreview />
+        <SocialLocationPreviewMap />
       </USMapWrapper>
     </ScreenshotWrapper>
   );
